refactor(server): extract coverage dump + codecov-ignore integration glue
The COVERAGE_DIR coverage dump only runs in instrumented builds (on shutdown),
so it can't lift the unit patch flag. Pull it out of server.ts into
utils/dump-coverage.ts so it can be codecov-ignored cleanly (server.ts keeps its
real coverage). Ignore the dump module, its merge script, and the extension
controller guards (install needs a live registry — no blackbox harness), the same
way controllers/auth-drivers were ignored before the blackbox flag.
